--- Sources/AllCache/CGSize+Util.swift
@@ -1,23 +1,0 @@
-//
-//  CGSize+Util.swift
-//  AllCache
-//
-//  Created by Juan Jose Arreola Simon on 5/13/17.
-//
-//
-
-import CoreGraphics
-
-func -(first: CGSize, second: CGSize) -> CGSize {
-    return CGSize(width: first.width - second.width, height: first.height - second.height)
-}
-
-extension CGSize {
-    var mid: CGPoint {
-        return CGPoint(x: width / 2.0, y: height / 2.0)
-    }
-    
-    var ratio: CGFloat {
-        return width / height
-    }
-}
