+++ Sources/AllCache/DiskCache.swift
@@ -2,31 +2,25 @@
 //  DiskCache.swift
 //  AllCache
 //
-//  Created by Juan Jose Arreola on 2/5/16.
-//  Copyright © 2016 Juanjo. All rights reserved.
+//  Created by JuanJo on 13/05/20.
 //
 
 import Foundation
 
-enum DiskCacheError: Error {
-    case invalidPath
-    case invalidData
-    case enumeratorError
-}
+private let diskQueue = DispatchQueue(label: "com.allcache.DiskQueue", attributes: .concurrent)
 
-public final class DiskCache<T> {
+public final class DiskCache<T, S: Serializer> where S.T == T {
     
     public let identifier: String
-    public let serializer: DataSerializer<T>
+    public let serializer: S
     public let cacheDirectory: URL
     
-    public var maxCapacity = 0
-    public internal(set) var size = 0
-    
-    private var shrinking = false
     private let fileManager = FileManager.default
+    public var maxCapacity = 0
+    public private(set) var size = 0
+    private var shrinking = false
     
-    required public init(identifier: String, serializer: DataSerializer<T>, directory: FileManager.SearchPathDirectory = .cachesDirectory) throws {
+    required public init(identifier: String, serializer: S, directory: FileManager.SearchPathDirectory = .cachesDirectory) throws {
         self.identifier = identifier
         self.serializer = serializer
         
@@ -36,34 +30,42 @@ public final class DiskCache<T> {
         if !fileManager.fileExists(atPath: cacheDirectory.path) {
             try fileManager.createDirectory(at: cacheDirectory, withIntermediateDirectories: false, attributes: nil)
         }
-        size = getCacheSize()
-        restrictSize()
     }
     
-    func getCacheSize() -> Int {
-        guard let enumerator = cacheDirectory.enumerator(includingPropertiesForKeys: [.totalFileAllocatedSizeKey]) else {
-            return 0
+    public func instance(forKey key: String) throws -> T? {
+        var result: T?
+        try diskQueue.sync {
+            let url = cacheDirectory.appendingPathComponent(validkey(from: key))
+            if fileManager.fileExists(atPath: url.path) {
+                result = try serializer.deserialize(Data(contentsOf: url))
+                diskQueue.async(flags: .barrier) {
+                    self.updateLastAccess(ofKey: key)
+                }
+            }
         }
-        return enumerator.compactMap { ($0 as? URL)?.totalFileAllocatedSize }.reduce(0, +)
+        return result
     }
     
-    public func object(forKey key: String) throws -> T? {
-        let url = cacheDirectory.appendingPathComponent(validkey(from: key))
-        if !fileManager.fileExists(at: url) {
-            return nil
+    public func removeInstance(forKey key: String) throws {
+        diskQueue.async(flags: .barrier) {
+            let url = self.cacheDirectory.appendingPathComponent(validkey(from: key))
+            let attributes = try? self.fileManager.attributesOfItem(atPath: url.path)
+            try? self.fileManager.removeItem(at: url)
+            if let fileSize = attributes?[.size] as? NSNumber {
+                self.size -= fileSize.intValue
+            }
         }
-        log.debug("🔑(\(key)) found on disk")
-        diskQueue.async { self.updateLastAccess(ofKey: key) }
-        
-        return try serializer.deserialize(data: Data(contentsOf: url))
     }
     
-    public func fileURL(forKey key: String) -> URL? {
-        let url = cacheDirectory.appendingPathComponent(validkey(from: key))
-        if fileManager.fileExists(at: url) {
-            return url
+    public func remove(olderThan limit: Date) {
+        diskQueue.async(flags: .barrier) {
+            let resourceKeys: [URLResourceKey] = [.contentAccessDateKey, .totalFileAllocatedSizeKey]
+            guard let enumerator = self.cacheDirectory.enumerator(includingPropertiesForKeys: resourceKeys) else {
+                return
+            }
+            let urls = enumerator.compactMap({ $0 as? URL }).filter({ ($0.contentAccessDate ?? limit) < limit })
+            urls.forEach({ self.removeIfPossible(url: $0) })
         }
-        return nil
     }
     
     public func allKeys() -> [String] {
@@ -74,84 +76,62 @@ public final class DiskCache<T> {
         })
     }
     
-    public func set(object: T, forKey key: String) throws {
-        let data = try serializer.serialize(object: object)
-        try set(data: data, forKey: key)
+    public func set(_ instance: T, forKey key: String) throws {
+        try set(data: try serializer.serialize(instance), forKey: key)
     }
     
     public func set(data: Data, forKey key: String) throws {
-        let url = cacheDirectory.appendingPathComponent(validkey(from: key))
-        try data.write(to: url, options: .atomicWrite)
-        log.debug("💽 Saved (\(key)): \(data.formattedSize)")
-        size += data.count
-        restrictSize()
-    }
-    
-    func updateLastAccess(ofKey key: String) {
-        let path = cacheDirectory.appendingPathComponent(validkey(from: key)).path
-        do {
-            try fileManager.setAttributes([.modificationDate: Date()], ofItemAtPath: path)
-        } catch {
-            log.error(error)
-        }
-    }
-    
-    public func removeObject(forKey key: String) throws {
-        let url = cacheDirectory.appendingPathComponent(validkey(from: key))
-        let attributes = try? fileManager.attributesOfItem(atPath: url.path)
-        if let fileSize = attributes?[.size] as? NSNumber {
-            size -= fileSize.intValue
-        }
-        try fileManager.removeItem(at: url)
-    }
-    
-    public func remove(olderThan limit: Date) {
-        diskQueue.async {
-            let resourceKeys: [URLResourceKey] = [.contentAccessDateKey, .totalFileAllocatedSizeKey]
-            guard let enumerator = self.cacheDirectory.enumerator(includingPropertiesForKeys: resourceKeys) else {
-                log.error(DiskCacheError.enumeratorError)
-                return
-            }
-            let urls = enumerator.compactMap({ $0 as? URL }).filter({ ($0.contentAccessDate ?? limit) < limit })
-            urls.forEach({ self.removeIfPossible(url: $0) })
+        diskQueue.async(flags: .barrier) {
+            let url = self.cacheDirectory.appendingPathComponent(validkey(from: key))
+            try? data.write(to: url, options: .atomicWrite)
+            self.size += data.count
+            self.restrictSize()
         }
     }
     
     public func clear() {
-        guard let enumerator = cacheDirectory.enumerator(includingPropertiesForKeys: nil) else { return }
-        for case let url as URL in enumerator {
-            removeIfPossible(url: url)
+        diskQueue.async(flags: .barrier) {
+            guard let enumerator = self.cacheDirectory.enumerator(includingPropertiesForKeys: nil) else { return }
+            for case let url as URL in enumerator {
+                self.removeIfPossible(url: url)
+            }
+            self.size = 0
         }
-        size = 0
     }
     
-    @inline(__always)
-    private func removeIfPossible(url: URL) {
-        do {
-            log.debug("💽 Deleting (\(url.lastPathComponent))")
-            try fileManager.removeItem(at: url)
-            size -= url.totalFileAllocatedSize ?? 0
-        } catch {
-            log.error(error)
+    func getCacheSize() -> Int {
+        guard let enumerator = cacheDirectory.enumerator(includingPropertiesForKeys: [.totalFileAllocatedSizeKey]) else {
+            return 0
         }
+        return enumerator.compactMap { ($0 as? URL)?.totalFileAllocatedSize }.reduce(0, +)
+    }
+    
+    func updateLastAccess(ofKey key: String) {
+        let path = cacheDirectory.appendingPathComponent(validkey(from: key)).path
+        try? fileManager.setAttributes([.modificationDate: Date()], ofItemAtPath: path)
     }
     
     func restrictSize() {
-        if maxCapacity <= 0 { return }
-        if size < maxCapacity { return }
+        guard maxCapacity > 0, size > maxCapacity else {
+            return
+        }
         diskQueue.async {
             if self.shrinking { return }
             self.shrinking = true
             do {
-                log.debug("💽 Original size: \(self.size)")
                 try self.restrictSize(percent: 0.8)
-                log.debug("💽 Final size: \(self.size)")
                 self.shrinking = false
             } catch {
-                log.error(error)
                 self.shrinking = false
             }
         }
+    }
+    
+    private func removeIfPossible(url: URL) {
+        do {
+            try fileManager.removeItem(at: url)
+            size -= url.totalFileAllocatedSize ?? 0
+        } catch {}
     }
     
     private func restrictSize(percent: Double) throws {
@@ -173,35 +153,33 @@ public final class DiskCache<T> {
 
 private let fileNameRegex = try! NSRegularExpression(pattern: "[/:;?*|']", options: [])
 
+@inline(__always)
 private func validkey(from key: String) -> String {
-    return "c" + fileNameRegex.stringByReplacingMatches(in: key, options: [], range: key.wholeNSRange, withTemplate: "")
+    let range = NSRange(location: 0, length: key.count)
+    return "_" + fileNameRegex.stringByReplacingMatches(in: key, options: [], range: range, withTemplate: "")
 }
 
 private extension URL {
     func enumerator(includingPropertiesForKeys keys: [URLResourceKey]?) -> FileManager.DirectoryEnumerator? {
-        return FileManager.default.enumerator(at: self, includingPropertiesForKeys: keys, options: [], errorHandler: {
-            (_, error) -> Bool in
-            log.error(error)
-            return true
-        })
+        return FileManager.default.enumerator(at: self, includingPropertiesForKeys: keys)
     }
 }
 
-extension Data {
-    var formattedSize: String {
-        switch count {
-        case ..<1024:
-            return "\(count) bytes"
-        case ..<(1024 * 1024):
-            return "\(Double(count) / 1024.0) Kb"
-        default:
-            return "\(Double(count) / (1024 * 1024)) Mb"
+extension URL {
+    var contentAccessDate: Date? {
+        return resource(forKey: .contentAccessDateKey) as? Date
+    }
+    
+    /// In bytes
+    var totalFileAllocatedSize: Int? {
+        return (resource(forKey: .totalFileAllocatedSizeKey) as? NSNumber)?.intValue
+    }
+    
+    @inline(__always)
+    func resource(forKey key: URLResourceKey) -> Any? {
+        if let values = try? resourceValues(forKeys: Set([key])) {
+            return values.allValues[key]
         }
-    }
-}
-
-extension FileManager {
-    func fileExists(at url: URL) -> Bool {
-        return fileExists(atPath: url.path)
+        return nil
     }
 }
