+++ Sources/ImageCache/UIButton+ImageCache.swift
@@ -9,7 +9,7 @@
 
 import UIKit
 import ShallowPromises
-import NewCache
+import AllCache
 
 public extension UIButton {
     
@@ -30,7 +30,7 @@ public extension UIButton {
         return ImageCache.shared.instance(for: descriptor).onSuccess { [weak self] image in
             self?.setImage(image, for: controlState)
             if let size = self?.bounds.size, originalSize != size {
-//                log.warn("Size mismatch, requested: \(originalSize) ≠ bounds: \(size) - \(self!.description)")
+                logger.error("Size mismatch, requested: \(originalSize) ≠ bounds: \(size) - \(self!.description)")
             }
         }.onError { [weak self] error in
             self?.setImage(placeholder, for: controlState)
